server_leapmotion/requirements_20250924.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
build==1.2.2.post1
certifi==2025.8.3
cffi==1.17.1
click==8.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
# Editable install with no version control (leap==0.0.1)
-e /home/fujii/leapc-python-bindings-main/leapc-python-api
leapc_cffi @ file:///home/fujii/leapc-python-bindings-main/leapc-cffi/dist/leapc_cffi-0.0.1.tar.gz#sha256=23537fc34cc593f79b592558c1aa15817e456930a83f526b1614197480f770cb
mcp==1.14.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyproject_hooks==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
rpds-py==0.27.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
